# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Loup — household services marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers:
    get:
      operationId: listProviders
      tags: [catalog]
      summary: List providers with optional filters
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Filter by category slug
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Free-text search over provider name and tagline
        - name: availableNow
          in: query
          required: false
          schema:
            type: boolean
          description: Only providers available right now
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
  /providers/{id}:
    get:
      operationId: getProvider
      tags: [catalog]
      summary: Provider profile with services
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Provider detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /services/{id}:
    get:
      operationId: getService
      tags: [catalog]
      summary: A single service with its provider — used to resolve a widget deep link (P1-7) to a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /providers/{id}/reviews:
    get:
      operationId: listProviderReviews
      tags: [catalog]
      summary: Reviews for a provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
  /reviews:
    post:
      operationId: createReview
      tags: [catalog]
      summary: Rate and review a completed booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Provider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        categorySlug:
          type: string
        categoryName:
          type: string
        tagline:
          type: string
        bio:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        jobsCompleted:
          type: integer
        yearsOnPlatform:
          type: integer
        verified:
          type: boolean
        availableNow:
          type: boolean
        responseMinutes:
          type: integer
          description: Typical response time in minutes
        startingPrice:
          type: number
        badges:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - categorySlug
        - categoryName
        - tagline
        - bio
        - rating
        - reviewCount
        - jobsCompleted
        - yearsOnPlatform
        - verified
        - availableNow
        - responseMinutes
        - startingPrice
        - badges
    Service:
      type: object
      properties:
        id:
          type: integer
        providerId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
      required:
        - id
        - providerId
        - name
        - description
        - price
        - durationMinutes
    ProviderDetail:
      allOf:
        - $ref: "#/components/schemas/Provider"
        - type: object
          properties:
            services:
              type: array
              items:
                $ref: "#/components/schemas/Service"
          required:
            - services
    ServiceDetail:
      allOf:
        - $ref: "#/components/schemas/Service"
        - type: object
          properties:
            providerName: { type: string }
          required: [providerName]
    Review:
      type: object
      properties:
        id:
          type: integer
        providerId:
          type: integer
        authorName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - providerId
        - authorName
        - rating
        - comment
        - createdAt
    ReviewInput:
      type: object
      properties:
        bookingId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required:
        - bookingId
        - rating
        - comment
...